rename email templates to message templates
