+++ baseapp-message-templates/MANIFEST.in
