+++ baseapp-message-templates/setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = baseapp_email_templates
+name = baseapp_message_templates
 version = 1.2
 description = A BaseApp app that allows the use of customizable e-mail templates.
 long_description = file: README.md
