+++ baseapp-message-templates/setup.py
@@ -4,7 +4,7 @@ with open("README.md", "r") as f:
     long_description = f.read()
 
 setup(
-    name="baseapp-email-templates",
+    name="baseapp-message-templates",
     packages=find_packages(),
     version="0.1",
     long_description=long_description,
