+++ baseapp-message-templates/testproject/requirements.txt
@@ -1,3 +1,3 @@
 # install "install_requires" from setup.cfg
 -e ./baseapp-core
--e ./baseapp-email-templates
+-e ./baseapp-message-templates
